Add OnAnswerCreated and OnQuestionBestAnswerChosen event handlers for notification management
