+++ src/domain/notification/application/subscribers/on-question-best-answer-chosen.ts
@@ -1,0 +1,40 @@
+import { DomainEvents } from '@/core/events/domain-events'
+import { EventHandler } from '@/core/events/event-handler'
+import { AnswersRepository } from '@/domain/forum/application/repositories/answer-repository'
+import { QuestionBestAnswerChosenEvent } from '@/domain/forum/enterprise/events/question-best-answer-chosen'
+import { SendNotificationUseCase } from '@/domain/notification/application/use-cases/send-notification'
+
+export class OnQuestionBestAnswerchosen implements EventHandler {
+  constructor(
+    private answersRepository: AnswersRepository,
+    private sendNotification: SendNotificationUseCase,
+  ) {
+    this.setupSubscriptions()
+  }
+
+  setupSubscriptions(): void {
+    DomainEvents.register(
+      this.sendQuestionBestAnswerNotification.bind(this),
+      QuestionBestAnswerChosenEvent.name,
+    )
+  }
+
+  private async sendQuestionBestAnswerNotification({
+    question,
+    bestAnswerId,
+  }: QuestionBestAnswerChosenEvent) {
+    const answer = await this.answersRepository.findById(
+      bestAnswerId.toString(),
+    )
+
+    if (!answer) {
+      return
+    }
+
+    await this.sendNotification.execute({
+      recipientId: answer.authorId.toString(),
+      title: `Your answer was chosen for "${question.title.substring(0, 40).concat('...')}"`,
+      content: `Your answer was chosen as the best answer for "${question.title} for the author"`,
+    })
+  }
+}
